refactor: domain에서 product와 stock패키지를 분리
