+++ src/main/java/com/woowa/woowakit/domain/product/domain/product/ProductName.java
@@ -1,4 +1,4 @@
-package com.woowa.woowakit.domain.product.domain;
+package com.woowa.woowakit.domain.product.domain.product;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
