+++ src/main/java/com/woowa/woowakit/domain/product/domain/product/ProductPrice.java
@@ -1,4 +1,4 @@
-package com.woowa.woowakit.domain.product.domain;
+package com.woowa.woowakit.domain.product.domain.product;
 
 import com.woowa.woowakit.domain.model.Money;
