+++ src/main/java/com/woowa/woowakit/domain/product/domain/product/ProductRepository.java
@@ -1,7 +1,7 @@
-package com.woowa.woowakit.domain.product.domain;
+package com.woowa.woowakit.domain.product.domain.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-	
+
 }
