+++ src/main/java/com/woowa/woowakit/domain/product/domain/product/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.woowa.woowakit.domain.product.domain;
+package com.woowa.woowakit.domain.product.domain.product;
 
 public enum ProductStatus {
