+++ src/main/java/com/woowa/woowakit/domain/product/domain/stock/Stock.java
@@ -1,0 +1,4 @@
+package com.woowa.woowakit.domain.product.domain.stock;
+
+public class Stock {
+}
