+++ src/test/java/com/woowa/woowakit/domain/product/domain/product/ProductQuantityTest.java
@@ -1,4 +1,4 @@
-package com.woowa.woowakit.domain.product.domain;
+package com.woowa.woowakit.domain.product.domain.product;
 
 import static org.assertj.core.api.Assertions.*;
